requirements.txt
py-clob-client>=0.17.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=12.0
aiohttp>=3.9.0
pytest>=7.4.0
sqlite-utils>=3.35
anthropic>=0.39.0
openai>=1.12.0
